backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy==1.26.4
websockets
opencv-contrib-python
pygame
keras
pillow
tensorflow
ultralytics
pyserial
kociemba
scipy
webcolors